more brfs cleanup
