+++ demo/package.json
@@ -47,7 +47,6 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-preset-stage-2": "^6.3.13",
-    "brfs-babel": "^1.0.0",
     "css-loader": "^0.26.1",
     "eslint": "^3.0.0",
     "eslint-config-uber-es2015": "^3.0.0",
