+++ examples/sample-layers/plot-layer/surface-vertex.glsl.js
@@ -18,18 +18,38 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#define SHADER_NAME trips-layer-fragment-shader
+export default `\
+#define SHADER_NAME graph-layer-vertex-shader
 
-#ifdef GL_ES
-precision highp float;
-#endif
+attribute vec4 positions;
+attribute vec4 colors;
+attribute vec3 pickingColors;
 
-varying float vTime;
+uniform vec3 modelCenter;
+uniform float modelScale;
+uniform float lightStrength;
+uniform float opacity;
+uniform float renderPickingBuffer;
+
 varying vec4 vColor;
+varying float shouldDiscard;
 
 void main(void) {
-  if (vTime > 1.0 || vTime < 0.0) {
-    discard;
-  }
-  gl_FragColor = vec4(vColor.rgb, vColor.a * vTime);
+
+  // fit into a unit cube that centers at [0, 0, 0]
+  vec4 position_modelspace = vec4((positions.xyz - modelCenter) * modelScale, 1.0);
+  gl_Position = project_to_clipspace(position_modelspace);
+
+  // cheap way to produce believable front-lit effect.
+  // Note: clipsspace depth is nonlinear and deltaZ depends on the near and far values
+  // when creating the perspective projection matrix.
+  vec4 center_clipspace = project_to_clipspace(vec4(0.0, 0.0, 0.0, 1.0));
+  float fadeFactor = 1.0 - (gl_Position.z - center_clipspace.z) * lightStrength;
+
+  vec4 color = vec4(colors.rgb * fadeFactor, colors.a * opacity) / 255.0;
+  vec4 pickingColor = vec4(pickingColors / 255.0, 1.0);
+
+  vColor = mix(color, pickingColor, renderPickingBuffer);
+  shouldDiscard = positions.w;
 }
+`;
