+++ examples/sample-layers/trips-layer/trips-layer-vertex.glsl.js
@@ -18,13 +18,27 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#define SHADER_NAME reflection-effect-vs
+export default `\
+#define SHADER_NAME trips-layer-vertex-shader
 
-attribute vec3 vertices;
+attribute vec3 positions;
+attribute vec3 colors;
 
-varying vec2 uv;
+uniform float opacity;
+uniform float currentTime;
+uniform float trailLength;
+
+varying float vTime;
+varying vec4 vColor;
 
 void main(void) {
-  uv = vertices.xy;
-  gl_Position = vec4(2. * vertices.xy - vec2(1., 1.), 1., 1.);
+  vec2 p = preproject(positions.xy);
+  // the magic de-flickering factor
+  vec4 shift = vec4(0., 0., mod(positions.z, trailLength) * 1e-4, 0.);
+
+  gl_Position = project(vec4(p, 1., 1.)) + shift;
+
+  vColor = vec4(colors / 255.0, opacity);
+  vTime = 1.0 - (currentTime - positions.z) / trailLength;
 }
+`;
