+++ src/experimental/effects/reflection-effect/reflection-effect-vertex.glsl.js
@@ -18,25 +18,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#define SHADER_NAME trips-layer-vertex-shader
+export default `\
+#define SHADER_NAME reflection-effect-vs
 
-attribute vec3 positions;
-attribute vec3 colors;
+attribute vec3 vertices;
 
-uniform float opacity;
-uniform float currentTime;
-uniform float trailLength;
-
-varying float vTime;
-varying vec4 vColor;
+varying vec2 uv;
 
 void main(void) {
-  vec2 p = preproject(positions.xy);
-  // the magic de-flickering factor
-  vec4 shift = vec4(0., 0., mod(positions.z, trailLength) * 1e-4, 0.);
-
-  gl_Position = project(vec4(p, 1., 1.)) + shift;
-
-  vColor = vec4(colors / 255.0, opacity);
-  vTime = 1.0 - (currentTime - positions.z) / trailLength;
+  uv = vertices.xy;
+  gl_Position = vec4(2. * vertices.xy - vec2(1., 1.), 1., 1.);
 }
+`;
