+++ test/fp64-test/webpack.config.test-fp64.js
@@ -32,11 +32,6 @@ module.exports = {
   module: {
     rules: [
       {
-        include: [resolve('../../src')],
-        loader: 'transform-loader',
-        options: 'brfs-babel'
-      },
-      {
         test: /\.glsl$/,
         use: 'raw-loader'
       }
