+++ webpack.config.js
@@ -36,25 +36,6 @@ module.exports = {
           this.value = content;
           return "module.exports = " + JSON.stringify(content); // eslint-disable-line
         }
-      },
-      // {
-      //   // Compile ES2015 using buble
-      //   test: /\.js$/,
-      //   loader: 'buble-loader',
-      //   include: [/src/],
-      //   options: {
-      //     objectAssign: 'Object.assign',
-      //     transforms: {
-      //       dangerousForOf: true,
-      //       modules: false
-      //     }
-      //   }
-      // },
-      {
-        // Inline shaders
-        include: [resolve('./src')],
-        loader: 'transform-loader',
-        options: 'brfs-babel'
       }
     ]
   },
